OCD-2104: Refactored chart app to use search results object
Two of the three charts don't need any more data than is already in
the search result object so can be generated without getting the
details of Listings. The third can be filtered to just the Listings
that have SED info before getting the details, cutting down the time
needed for that chart by about half currently.