test: fix unit tests
Eclipse reported a potential leak with the context not being closed.
Closing that context in the @BeforeClass had it closed before it could
be used in the tests. It can't be closed in an @AfterClass because
tests then fail depending on the order in which they're run. This
commit undoes the previous "fix", though it leaves the Eclipse error
still around. No fix for that is immediately apparent.