feat: Re-enable the Developer Attestation Check-In Report
There was common functionality being used in the check-in report and the
regular Developer Attestation Report - in fact, the check-in report was
calling the data collector for the attesation report which seemed like a
lot of extra work (running both reports essentially). I refactored the
shared functionality into the DeveloperAttestationPeriodCalculator class.
More calculations could occur there but for now, there is just 1.

[#OCD-3985]