feat: Create a developer search service
The developer search service allows a user to pass in a set of search
parameters and receive a set of matching developer search results.
A search result is not a complete developer object, but instead contains
a subset of developer fields.

For search purposes, a developer is associated with ACBs from all of its
listings regardless of listing status.

[OCD-4017]