feat!: Include removed criteria in Cures Update calculation
The CuresUpdateService was not considering removed criteria when
calculating whether a listing was Cures Update version or not. After
ERD-Phase-2 when we mark the "old" criteria as removed we still need to
use the to determine a listing's Cures Update status.

Also, remove placeholder code that checks if 24 months have passed since
cures Rule Effective Date.

[#OCD-4051]