feat!: Do not save developer activity if no data has changed
If a request to update a developer object does not contain the addressId
or contactId fields, new address/contact will be created which will have
a different ID than the previous, tricking the system into thinking
there was an update to data when it was really just an ID change. To fix
this, we will just compare all address and contact values (counting null
and "" as the same) and not do any updates if there are no changes.

[#OCD-3899]