refactor: Separate UCD Process domain obj from UCD Process on a listing
There was a single UcdProcess object used both to get UCD
Process domain objects from the database and as the set of UCD Processes
associated with criteria on a listing. In order to eventually have CRUD
operations on the UCD Process domain I wanted to separate these two
distinct types of UCD Process objects.

[#OCD-4039]