fix: Add email sending to upload job for processing errors
An email has always been sent from the ListingUploadController, and
still gets sent from there in the event that any required headings are
missing and we cannot proceed with the file. However, after moving the
initial upload validation to a quartz job we lost the feature of sending
an email when there was any sort of error parsing, normalizing or
validating the file.

[#OCD-3263]