feat: Use listing normalizer to populate fields on update
There was previously a method to "sanitize" listing data on update which
was filling in a few choice IDs that were sometimes missing when the UI
sent listing JSON for an update. The normalizer classes that have been
written and used during flexible upload fill in possibly missing IDs on
many additional fields and have many unit tests validating their
behavior. It was always the goal to move to using these normalizers
during update and now is the time. The "sanitize" code is being removed
in favor of normalizers.

Most relevant to this ticket, the normalizers fill in missing CQM IDs.
If a user switches from c3-original to c3-cures criteria when editing a
listing, the UI is not passing back a specific C3 in the CQM data so the
normalizer can figure that out based on which one is attested.

[#OCD-4071]