feat: Add "child" processing to ICS Normalizer
If a listing includes any children in its ICS declarations, the
normalizer will fill in any missing IDs for those children.

The UI does not allow us to update children (only parents) so it is
unlikely this code will ever get called. It was added as a precaution
because it was part of the newly removed "sanitize" method in
CertifiedProductManager.

[#OCD-4071]