feat: Use internal error email address for all such errors
We had Upload Errors, Developer Split Errors, and Developer Merge Errors
all potentially going to separate email addresses AKA Teams channels.
This commit combines all of those types of errors to go to the same
place and will allow us to consolidate some of our Teams
channels/simplify some code.

In addition, we will send a message to the internal error channel if we
detect an email trying to be sent without a recipient.