fix: Ignore UCD Processes without actual process
A user of the API can theoretically send in a UCD Process associated
with a listing that has only a criteria and no UCD Process itself. The
normalizer should just remove that - it is sort of hopeless. We cannot
find an ID associated with it. All other validation checks still apply.

[#OCD-4039]