fix: Use a single instance of GraphServiceClient
I noticed some socket timeout errors in the logs after this code was on
QA for a while. Not every time and not very consistently. But googling
the error turned up a recommendation that there only be a single
GraphServiceClient per application. Previously there was one per
SendEmailJob. But I think the GSC keeps a connection pool so if you end
up instantiating a bunch of GSC's you'll eventually run out of
connections/sockets/whatever. My theory is that is what was happening on
QA.

[#OCD-3970]