feat: Use HttpClient instead of AsyncHttpClient for URL Calls
The thought was that the AsyncHttpClient library hasn't been updated in
over a year since sometime in 2021. I wanted to try changing the library
used to be one that is at least regularly updated. We cannot use the
latest HttpClient v5 without also upgrading our version of Spring so we
are just trying HttpClient v4 here.

[#OCD-4081]