feat!: Add property to configure whether emails are deleted from sent
We originally thought we could do something neat by clearing out emails
sent from non-prod systems, but it seemed that sometimes the emails were
deleted too quickly (like before they were sent). Allowing separate
toggling of that behavior means we can turn it off and hopefully not
lose any emails. If we want to try to get that behavior to work better
in the future, we will be able to turn it back on.

The property name is emailBuilder.config.saveToSentItems

[#OCD-4102]