fix: Add normalization to listing validation report
We added normalization to the work of updating a listing (which is done
before validation) but did not add the same logic to the listing
validation report. This resulted in a few listings giving errors on
update due to the normalization code running that never appeared in the
listing validation report.

[#OCD-4111]