fix!: Correctly handle SED task/participants on listing edit and upload
An issue was discovered when a user tried to edit a listing and they
were not able to add new test tasks (along with their participants). It
was all related to the use of a UI-generated negative number unique ID.
We need this uniqueId to be present and utilized because the same test
participant can be used throughout multiple test tasks within the same
listing.

Working on this bit of code revealed another bug that had existed since
we turned on the flexible upload. Any listing uploaded since then with
SED has way too many test participants - we were not deduping them based
on the uniqueId.

[#OCD-4117]