feat!: Add /qms-standard endpoints supporting CRUD operations
This commit also deprecates the /data/qms_standards endpoint and removes
as much fuzzy matching code as possible. The entity is removed and the
table is no longer referenced. Existing deprecated fuzzy endpoints will
be no-ops until the endpoints can be removed.

The QMS Standard code is refactored into a qmsStandard-specific package.

[#OCD-4041]