fix: Use the vanilla activity date for listing questionable activity
Listing Questionable Activity was using a "new Date()" when an entry
into the table was being created. This made the questionable activity
date just slightly different (milliseconds) from the Activity date. When
searching through Activity to see if there was already a Questionable
Activity entry for that, the dates really need to be identical. I think
it is true anyway that we want the activity_date columns of the
"activity" and "questionable_activity" tables to match exactly.

This previously was a feature (and remains a feature for the other
non-listing types of questionable activity... developer, product,
version). I think it got accidentally changed to create different dates
with a refactor of Listing Questionable Activity in ticket OCD-3748.

[#OCD-4086]