feat: Remove onetime job to update measures
I think we were keeping the job around to possibly make small updates to
and use again in the future, but it had a lot of code to deal with
legacy measure mappings and was using the legacy measure names. My
thought is that if we need to have a job remove measures again it will
be very different from this one which is why I decided to delete it
instead of update it.

[#OCD-3996]