fix: Re-instate combining of measures to avoid duplicates
Combining measures in the MeasureNormalizer allows measures to be
specified in multiple lines in the upload file but then combined into
one measure + associated criteria during normalization. This was
existing behavior (sort of a side effect of combing measures at the
criteria level with measures at the listing level) but it doesn't make
sense to remove that existing functionality for users.

[#OCD-3996]