fix: Clear ICS relationships from shared store before listing update
The ICS relationships need to be cleared both before and after
because if an ICS relationship is REMOVED from the listing, it is no
longer available when the aspect/annotation runs and misses being
cleared out. It will still show an ICS relationship in it's listing
details when none exists.

[#OCD-3910]