feat: Remove derivable columns from ICS Errors Report Creation
Going forward, we plan to treat all "reporting" tables in this way. We
will do a regular delete of the data and insert all of the new data at
once. There is no need to keep derivable columns in the reporting tables
if they can be gotten efficiently from listings via the shared store.

[#OCD-4107]