feat: Remove Ashwini from the Trigger Developer Ban job
Ashwini will get removed in any env when we reset-quartz jobs. If/when
we need to test the trigger developer ban feature, we will have to add a
user to this job. But her email should not be in the codebase and
probably should not be in production (or any environment) by default.

[#OCD-4060]