fix: Validate that developer has other products when transfer w/o hist
There was validation added to check that the old developer product owner
has other products when product ownership is changed with an entry in
the history. However, if there is no entry in the history (user clicks
'x' in the UI to remove the automatically added history entry) then we
were allowing the ownership change regardless. This commit adds code to
validate the "developer must have at least one other product" rule in
the event of an ownership-without-history-change.

[#OCD-4050]