fix: Allow developer to lose all products during developer merge
We created a rule that developers could not lose all of their products.
However, applying this rule on developer merge makes developer merge not
work (because by definition we are removing all the products of a
developer!).

I don't love the solution I came up with here but I thought about it for
a while and did not see a better way. I am settling for it since we
likely will be re-writing this workflow relatively soon with a ticket in
the backlog that is going to let developers transfer ownership of ALL
products from one to another.

[#OCD-4050]