fix: Do not consider "deleted" developers for "must have a product" rule
When merging developers, the flow is to transfer all products from both
developers to a newly created developer. Then mark the merged developers
as deleted. If a person later goes to edit ownership history of one of
the moved products and edits it in a certain way, they get an error
related to the "deleted" developer. This commit adds code to not care
about whether deleted developers have any products or not.

[#OCD-4050]