fix!: Handle null developer contact name in Developer Access Report
We got into the state where a developer had null contact info but was
still included in the access report when we changed some product
ownership. If a product that has an active listing is transferred to a
developer that only has retired listings, then that developer now must
be included in the report. The developer previously was never required
to have contact information and we did not require it when transferring
the product. Maybe we should?

[#OCD-4050]