fix: Include CAP Date fields plus capStatus in JSON
It is more desirable to ignore these fields in the JSON when considering
only the UI. We don't want to use them for display of anything. However,
if they are ignored in the JSON then they do not get written to the
shared store and when a listing is pulled back out of the shared store
it is missing these fields - CAP Status and the provided dates - which
are required to determine capApproval, capEnd, and capMustComplete.

The only way to handle this, I believe, is to NOT ignore the fields when
the objects go to/from JSON. They will be present for UI purposes but
should most likely never be needed.

[#OCD-4155]