feat!: Add complaint search filters for listing ID, surv ID, criteria ID
The UI was doing a filter by surveillance ID using an endpoint that was
being deleted and had been deprecated for a while. I needed to add
filtering by surveillance ID to replace that functionality and it made
sense to add listing ID and criteria ID filtering as well since those
are the other things that can be associated with a complaint.

[#OCD-4138]