fix!: Handle "Jira Unavailable" better in Listing Search Service
The /search/v* endpoints returned a 400 error if a user requested
compliance filtering when Jira was unavailable but that was not a good
strategy. Instead we will return 0 results if the user sends in
compliance filters. The API always returns a field letting a consumer
know that direct reviews are unavailable.

[#OCD-4183]