refactor: Change CQM version Set to LinkedHashSet
The LinkedHashSet preserves order. Alternatively we could have changed
these to be lists but I dont know if there is a good reason they were
made Sets to begin with.

When built from scratch, the allVersions and successVersions were
correctly ordered but when read from the Shared Store, they were getting
unordered so rather than use the generic Set/HashSet, I switched it to
LinkedHashSet because that one has a guaranteed iteration order.

[#OCD-4064]