feat: Do not check for questActivity if no activity is added
We were checking for questionable activity even when no activity is
added (like if the original and updated listing JSON is identical so
there were no changed). So we could have a situation where there is NO
activity but there IS a questionable activity for something like
2011/2014 listing edited. I think that it doesn't make sense to check
that something is questionable if there is no activity in the first
place.

[#OCD-4064]