feat!: Improve g3Sed download file creation job
The job was using some older code and holding all the listing details in
memory (for listings that attest to g3) and then dumping them out to a
file. On PROD, this sometimes makes our server run out of memory.

This commit updates the job to use our newer file-writing strategy and
streams individual listing data to the file.

[#OCD-4180]