feat: Use ActivityExclude to ignore listing fields when writing activity
No need to include error/warning messages when writing listing activity.
If we include them then we need to sort them and that has become messy
and tricky with the addition of data and business error fields.

We may decide to ignore additional fields that are currently included
but don't really describe "a listing".

[#OCD-4064]