refactor: Slightly improve performance of the check-in report
Gathering the information about the most recent attestation change
request for the current attestation period for each developer was
performing linearly and would increase over time each quarter when more
change requests are added for each developer. Locally it was taking
about 3 seconds per developer and around 20 minutes total to run the
report and would slowly increase over time. This update runs in about
15 minutes locally and I don't *believe* it will increase over time.

The previous code was getting all attestation change requests for a
developer, then getting the details of each of those change requests in
order to sort them by submission date. The second step would be
increasing in time as more attestation periods go by. I changed this to
make a query to get the ID of the most recent attestation change request
for the developer and then a second query to get the details of that
change request. I think in the long term this will perform better.

[#OCD-4147]