feat: Specify whether to populate dependent objs when getting change req
This is a performance improvement. We were getting the whole form and
frequently also the warning messages for the allowed responses which is
a lot of roundtrips to Redis. Where this code was being called though we
really only needed the CR status and type, not any of that other stuff.
So I just added a boolean to indicate whether to go fetch it or not.

[#OCD-4200]