refactor: Add checks for null edition throughout
The validator/reviewer that ensures an edition is non-null is still in
place for now. I just updated everywhere I could find in the code that
was using a field of the edition object to make sure that it was not
null. We won't be able to test this without a listing that actually has
a null edition.

[#OCD-4269]