refactor: Consolidate to one ConformanceMethod class
We sometimes need to return Conformance Methods with criteria, such as
when loading the Management page or building a screen to edit a listing
and get all the conformance methods available, and other times we need
to return Conformance Methods without their allowed criteria, such as
when building a listing details object.

I had initially done this with two different classes but that is not a
good solution because all the other property names are in two places,
etc. Instead, we have one ConformanceMethod class with a "criteria"
property annotated to be left out of the serialized JSON when it's null.
The ConformanceMethodEntity class has two different "toDomain"-type
methods: one that expects criteria to have been fetched in the HQL and
fills them in, and one that does not. We will use the former when
getting conformance methods for managmenent purposes and the latter when
building listing details.

[#OCD-4129]