feat: Include editionless listings in some statistics
This was in an effort to include the new criteria in statistics, and
editionless listing may attest to the new criteria. The one statistics
that is not obvious how to include is the count of listing statuses per
developer/product for each edition. I feel like we could drop the
edition entirely but that seems like a different ticket.

[#OCD-4336]