feat: Do not require entire listing to be serialized for DevBanJob
It is really really hard to test that every single field of every single
listing is Serializable without putting data in every field and trying
to serialize it. So to future-proof the Trigger Developer Ban Job a
little bit we just won't send in the whole listing details. We can send
in the listing ID only and then get it's details during the run of the
job.

[#OCD-4420]