feat!: Deprecate GET /data/developers, /data/products
These two endpoints returned a list of developer or product names along
with a set of how many listings were in each possible status for that
developer or product. Allowing a certification status to have a future
date creates added complexity in the view that supplies in the data, and
upon a close inspection it turns out the CHPL application is not using
these two endpoints or the data they present in any way. So they can be
removed.

[#OCD-4452]