feat: Only allow one update status job per listing per day
The update current certification status job includes an "activityId"
passed into the job. That activityId is inserted as an observation to be
handled when subscription notification emails go out. If an ACB saved a
listing with tomorrow's status as 'Withdrawn by ONC-ACB' then a job
would be scheduled to run tomorrow and handle that status change using
the activity from the "Save" action. If they realized they made a
mistake and went back in and set a the status to actually be 'Withdrawn
by Developer' tomorrow, then we want to cancel the job scheduled for
'tomorrow' referencing the 'Withdrawn by ONC-ACB' activity and
reschedule a new job for tomorrow referencing the 'Withdrawn by
Developer' activity. Otherwise notifications could go out to subscribers
saying the listing had been both 'Withdrawn by ONC-ACB' AND 'Withdrawn
by Developer' when really the first status change never happened.

[#OCD-4452]