feat!: Accept educationType and age objects for test participants
The Angular UI is not changing and will still use ageRangeId, ageRange,
educationTypeId, educationTypeName. These fields are marked as
deprecated and will be removed when listing edit is turned off in the
Angular UI.

In addition, the Angular UI sets a test task or participant ID to a
negative number when it is newly created in the UI. The API code to
handle this cannot be removed until Angular UI editing is turned off, so
this code remains. But we are now also supporting a "null ID and
non-null uniqueId" combination to indicate a new test task or
participant. This will be the model kept going forward with the React UI
editing.

[#OCD-2838]