fix: Force validation of roles when going to a 'by ONC' status
We don't run any validation if a listing is being put into a non-active
status. This means we weren't validating if the user performing the
action was role admin/onc if the listing was being put in Suspended by
ONC or Terminated by ONC. And we need to do that validation always.

[#OCD-4452]