refactor: Remove all quarterly report DTOs
As of this commit, there is still some work to do to untangle the
RelevantListing and PrivilegedSurveillance objects. The complications
come in because the same single Surveillance could have different
"privileged" values (those that get put in the report) from quarter to
quarter. The mapping was done in the DTOs that were removed and I need
to figure out how to put that back in without breaking backwards
compatibility. The code compiles now though.

[#OCD-4484]