fix: Normalize standards based on cert day/today as appropriate
The rules for reviewing standards apply differently during upload (using
certification day) than during edit (using current day). This also
changes the normalization of standards to be applied in the same way.
When uploading a listing, it is normalized per certification day and
when updating an existing listing it is normalized per the current day.

[#OCD-4451]