fix: Also declare exception thrown for surveillance delete
The UI does not let a user click the "Delete" button for a surveillance
unless a reason has been entered, so while this probably would have had
the same issue as the listing update w/missing reason there is not a
good way to test it. If a user had tried to delete a surveillance using
the API and not provided a reason I think they would have gotten the
same 500 error we saw with listings but really no one does this.

[#OCD-4576]