fix: Show expected error message if product owner is banned
The prior code showed a helpful error message and something about the
changes in this PR caught the error earlier in the Spring
PermissionsEvaluator classes and was just returning AccessDenied. I
don't really like the change I've done in this commit but I don't
understand how it was NOT returning just Access Denied before or how the
other changes I made resulted in it starting to return that. It should
have always been giving Access Denied. Anyway, the change here results
in the desired behavior for the user so I guess it's fine.

[#OCD-4554]