fix: Add transactional to DAO method
This method is called from a Permissions class to check whether the
current user can update product ownership. For some reason, in the
Tomcat 9/javax world, this works fine, but in the Tomcat 10/jakarta
world the permissions class does not have a transaction. Annotating the
method works to fix it but I'm not sure if it's the best permanent
solution.

[#OCD-4554]