refactor: Remove deprecated Hibernate Where annotation usage
The @Where annotation is deprecated in this newer version of Hibernate
to which we have recently upgraded. There are a few options when
replacing @Where (which we were using entirely as a shortcut for not
returning soft-deleted items). I tried using the new @SoftDelete
annotation on the subscription entities, but had a hard time getting
that working properly. The change also spiders into all entities because
if you use the @SoftDelete annotation, you cannot declare a "deleted"
column in the Entity, but we are declaring that via a superclass used
among ALL entities called EntityAudit.

I don't find the HQL to be all that much worse with a few "and deleted =
false" statements around, but we could revisit this in the future if and
when we remove the @Where annotation on other entities.

[#OCD-4483]