feat!: Deprecate 'uniqueId', add 'friendlyId' to tasks and participants
Save the friendly id on upload of a new listing. Needs some coordination
with UI because we were (and still need to?) use the uniqueId field as a
negative number when a user edits a listing to add a new test task.
Maybe we should just require them to enter a friendlyId with new test
tasks in the UI and stop using the uniqueId field.

[#OCD-4636]