fix: Specially format text values that might look like numbers
This includes all of the test task calculated data (standard deviations,
means, etc), and anything that might look like a version (where it's
often something like "1.0"). Exporting as a CSV and immediately
re-uploading works fine without this commit, but if someone opens that
CSV in Excel to make edits and re-saves it, all of these types of values
may be "helpfully" converted by Excel to numbers rather than treated as
text. So "1.0" goes to "1" and shows up in the diff when this happens,
and we want to avoid that.

[#OCD-4682]