feat: Make user activity metadata accessible only to ONC/ADMIN
There was code that theoretically allowed ACB and CMS users to see
activity for users that they had access to (i.e. a new Drummond ACB user
was created then other Drummond users should be able to see that). But
that was not going to work when we move to Cognito, and I don't think we
ever made use of it anyway. So this simplifies the code and keeps access
to those who might actually use it.

[#OCD-4727]