feat: Leave non-editable removed criteria out of CSV if user is ACB
When reviewing the checklist of things that should happen when a9 gets
removed, I realized that certain criteria are considered "editable" -
that is if they have been removed less than 1 year ago. I was including
ALL criteria in the listing CSV File but I think that might be confusing
for users actually because they would see some "original" criteria that
actually could not be added anyway. This commit changes the criteria
included in the download file - only those that are 1) attested or 2)
editable to the user are included. If ADMIN/ONC gets the CSV file then
they do get all the original criteria because they can do pretty much
whatever they want.

[#OCD-4705]