feat!: Deprecate fields in dev, prod, listing activity metadata
The additional fields being filled in for developer, product, and
listing activity metadata are not going to be used any longer. There
were issues parsing this data out because of work done in OCD-4554
(repurposing a field called 'statuses' which exists in old activity and
gives errors when we tried to parse that old activity).

We are basically just no longer doing work we didn't need to be doing in
the first place, now that the individual Activity pages are going away.

[#OCD-4740]