feat!: Deprecate surveillance endpoints, certifiedProduct field
The surveillance PUT, POST, and DELETE endpoints are being moved to the
certified product controller. The new endpoint will have the format
/certified-products/<id>/surveillance/<id>. Since the certified product
ID will be part of the URL path, it does not need to be included as part
of the Surveillance object. That design was always a little weird. This
commit also removes use of the certifiedProduct field as much as
possible.

[#OCD-3928]