fix!: Remove invalid SVAPs during listing edit with a warning
As part of the effort to consolidate new listing reviewers from edit
listing reviewers, this commit refactors the SVAP reviewers. There was a
reviewer for new listings and reviewer for editing listings but they
were treating invalid SVAPs slightly differently, with the "new listing"
reviewer having the preferred functionality.

Upon new listing or edit listing, invalid SVAPs should be removed for
the user and a warning should be presented to them. Other behavior is
slightly different between new/editing and new listings are allowed to
have removed SVAPs as long as they have ICS. Edited listings may keep
any removed SVAPs already present but should be presented a warning in
the event that they add additional removed SVAPs as part of their edit
regardless of the listing's ICS status.

[#OCD-4834]