fix: Use File setReadable method
The PosixFileAttribute fine-grained file permissions did not result in
the "global" read bit being set despite code being written to do this.
Research indicates the following reason for that not working:

When the access permissions are set at file creation time then the
actual value of the permissions may differ that the value of the
attribute object. The reasons for this are implementation specific. On
UNIX systems, for example, a process has a umask that impacts the
permission bits of newly created files. Where an implementation supports
the setting of the access permissions, and the underlying file system
supports access permissions, then it is required that the value of the
actual access permissions will be equal or less than the value of the
attribute provided to the createFile or createDirectory methods. In
other words, the file may be more secure than requested.

[#OCD-4842]