feat: Improve error messages about missing dependent criteria
The error message were originally written as "you have an (a)(*)
criteria so you need to also have (d)(1)", for example. This worked for
a while when it truly applied to all a-criteria (hence the "*"). This
has not been the case for some time, so I wanted to update the error
messages to be specific - "you have (a)(2) so you also need to have
(d)(1)".

[#OCD-4838]