fix: Always use MAX_PAGE_SIZE for search results
When making a query with quite a lot of results (for example, getting
all withdrawn listings has > 1000), it takes a lot of time to page
through them if the pages are small. The Power BI reports were
frequently failing on the withdrawn-listings call. The ListingStatistics
report, which makes this call, makes 7 calls to the CHPL API which
appear to be batched in groups of 5. Each call appears to timeout after
30 seconds. The withdrawn-listings call was almost always taking more
than 30 seconds. If any one of those 7 calls to refresh the report fails
due to timeout, then all the calls are tried again. This can go on for
at least 4 attempts to totally refresh the report. Increating the page
size requested dramatically reduces the load time of withdrawn-listings
(and all other listing searches that this commit touched).

[#OCD-4807]