fix: Use springdoc 2.7 in order to keep openapi version 3.0.1
Using the latest springdoc, 2.8.8, makes the generated swagger
documentation use openapi spec version 3.1. The UI component DOES NOT
work with that. It ONLY works with version 3.0. We cannot upgrade the UI
component until the React conversion is completed and then maybe even
updated to a newer version of React. I am worried about the next time we
have to upgrade spring/spring-boot/springdoc (in one year) because if we
have to be able to support the openapi 3.1 spec we need the UI component
to be upgraded by then.

[#OCD-4898]