feat!: Update validation for new "gap" conformance method
This commit removes the dependency on conformance method names
("Attestation" was hard-coded) in favor of id-based lookups that we use
for criteria and other things. Both Attestation and Gap conformance
methods do not require a version. Attestation is actually not allowed to
have a version at all, but I think "Gap" is.

[#OCD-4978]