fix: Create download file dir on demand if it does not exist
This is not needed for our running systems, but was noticed as an issue
with the tomcat docker container that runs the rest services. The
download directory didn't exist on there. This commit updates the code
to always get the download dir from one place and to create it if it
does not exist.

[#OCD-5114]