fix: Remove Builder.Default from JWTAuthenticatedUser
I am not sure why this was giving a compilation error. The same
annotation is used all over the place without issue. But it was not a
big deal to take it away and make sure the list of authorities that was
being initialized is actually initialized in the few places where it is
built.

[#OCD-4907]